Update and rename AutoAob.cs to core.cs
Updates:

Some try statements
Typpo fixing
Add the possibility to add a int to the address
